# alodo-backend/public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "ALODO — API du diagnostic MPME",
        "description": "Documentation générée depuis les annotations des contrôleurs. Obtenir data.token via login/register, puis utiliser Authorize. Les identifiants des exemples sont fictifs.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Backend courant"
        }
    ],
    "paths": {
...
        "/api/display/questionnaire": {
            "get": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Charger le questionnaire et les réponses",
                "description": "DiagnosticController::indexQuestionnaire. Rôle user. Domaines non vides triés par ordre, questions par ordre puis id. Retourne les réponses du diagnostic connecté. Toutes les questions, y compris Général non noté, sont obligatoires à la finalisation. Un diagnostic completed reste consultable.",
                "operationId": "indexQuestionnaire",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Questionnaire"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/store/answers": {
            "post": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Enregistrer ou terminer le diagnostic",
                "description": "DiagnosticController::storeAnswers. Rôle user, propriétaire seulement. pending sauvegarde les réponses envoyées sans supprimer les autres. completed revalide toutes les réponses, calcule et persiste le résultat, termine le diagnostic puis déclenche le job IA. La finalisation est définitive. result=null pour pending. Pour completed, lire analysis_status : l’IA peut être en attente, en cours, terminée ou en échec. Un score global null est possible même avec toutes les réponses. Pas de tableau responses vide, même pour completed. Le barème actuel requiert les 12 questions configurées (texte/choix unique), malgré les 4 types acceptés par la validation des réponses.",
                "operationId": "storeAnswers",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnswersRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AnswersData"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/Validation"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/diagnostics/{diagnostic}/result": {
            "get": {
                "tags": [
                    "Résultats"
                ],
                "summary": "Consulter le résultat persisté",
                "description": "DiagnosticController::showResult. Rôle user, propriétaire seulement. 404 si aucun résultat. Le score est une chaîne décimale ou null ; scoring_details contient des nombres. Au moins 2 questions évaluables par domaine et les 3 domaines évaluables sont requis pour le score global. Général n’est pas noté. Lire régulièrement pendant analysis_status=pending/processing ; arrêter à completed/failed. Aucun appel IA n’est déclenché par cette lecture.",
                "operationId": "showResult",
                "parameters": [
                    {
                        "name": "diagnostic",
                        "in": "path",
                        "description": "ID du diagnostic appartenant au compte connecté, obtenu via le questionnaire. Ce n’est pas l’ID du résultat.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Result"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/diagnostics/{diagnostic}/interpretation/retry": {
            "post": {
                "tags": [
                    "Résultats"
                ],
                "summary": "Relancer une interprétation échouée",
                "description": "DiagnosticController::retryInterpretation. Rôle user, propriétaire seulement. Aucun corps nécessaire. Accepte uniquement analysis_status=failed avec les données nécessaires. Ne recalcule pas le score. Le 202 confirme la relance, pas la réussite de l’IA ; relire le résultat. Le job nécessite un worker avec une connexion de queue asynchrone.",
                "operationId": "retryInterpretation",
                "parameters": [
                    {
                        "name": "diagnostic",
                        "in": "path",
                        "description": "ID du diagnostic appartenant au compte connecté, obtenu via le questionnaire. Ce n’est pas l’ID du résultat.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Result"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Questionnaire": {
                "required": [
                    "diagnostic",
                    "domains"
                ],
                "properties": {
                    "diagnostic": {
                        "required": [
                            "id",
                            "status",
                            "completed_at"
                        ],
                        "properties": {
                            "id": {
                                "type": "integer",
                                "minimum": 1,
                                "example": 1
                            },
                            "status": {
                                "type": "string",
                                "enum": [
                                    "baseline",
                                    "pending",
                                    "completed"
                                ]
                            },
                            "completed_at": {
                                "description": "Date brute du modèle Diagnostic, sans cast datetime ; null avant finalisation.",
                                "type": "string",
                                "example": "2026-09-14 10:00:00",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "domains": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QuestionnaireDomain"
                        }
                    }
                },
                "type": "object"
            },
...
            "AnswersRequest": {
                "required": [
                    "diagnostic_id",
                    "status",
                    "responses"
                ],
                "properties": {
                    "diagnostic_id": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed"
                        ]
                    },
                    "responses": {
                        "description": "Chaque question_id doit être distinct. Les réponses omises restent enregistrées.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnswerInput"
                        },
                        "minItems": 1
                    }
                },
                "type": "object"
            },
...
            "AnswersData": {
                "required": [
                    "result",
                    "responses",
                    "newStatus",
                    "completed_at"
                ],
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/NullableResult"
                    },
                    "responses": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SavedAnswer"
                        }
                    },
                    "newStatus": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed"
                        ]
                    },
                    "completed_at": {
                        "description": "Date de finalisation ; format brut du modèle, non garanti ISO 8601 dans cette réponse.",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "Unauthorized": {
                "description": "401 : token absent/invalide ou identifiants de connexion incorrects.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "403 : rôle non autorisé.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "404 : ressource absente, diagnostic inaccessible ou rôle non configuré.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Validation": {
                "description": "422 : données invalides. Les clés responses.{index}.valeur concernent la soumission ; answers.{question_id} concerne la complétude finale. diagnostic_id peut signaler un diagnostic déjà terminé.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "Conflict": {
                "description": "409 : seule une interprétation failed avec ses données de calcul et son instantané peut être relancée.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "ServerError": {
                "description": "500 : erreur serveur, notamment incohérence entre questions et barème lors de la finalisation. La transaction de cette soumission est annulée.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Laravel Sanctum, sans préfixe Bearer dans Authorize.",
                "scheme": "bearer"
            }
        }
    },
...
}